requirements.txt
pydantic-ai
google-cloud-bigquery
logfire